optimize and simplify eval()
the previous method used a generic strategy which allowed us to
evaluate any fragment by appending it to the end of the bytecode.
however, this ran into complications because of dead code elimination
during assembly. plus the strategy was only used for `eval()`, for which
a much simpler compilation strategy is to simply prepend the fragment to
be assembled and re-assemble the entire program.