fix[perf]: fix performance issue in computing source map (#392)
the source map is not shared between VyperContract instances since it is
not cached on CompilerData. this computes it before putting it into the
disk cache.

instead of computing the source map in VyperContract, compute it in
boa/interpret.py -- and invalidate the cache if we didn't find the
source map.